Use RelatedWorks from internal_model
